+++ src/Templateless/Components/Otp.cs
@@ -1,5 +1,4 @@
 using Newtonsoft.Json;
-using Templateless.Components;
 
 namespace Templateless.Components
 {