+++ src/Templateless/Components/ViewInBrowser.cs
@@ -7,10 +7,10 @@ namespace Templateless.Components
 		[JsonProperty("id")]
 		public ComponentId Id { get; private set; }
 
-		[JsonProperty("text")]
-		public string Text { get; private set; }
+		[JsonProperty("text", NullValueHandling = NullValueHandling.Ignore)]
+		public string? Text { get; private set; }
 
-		public ViewInBrowser(string text = "")
+		public ViewInBrowser(string? text = null)
 		{
 			Id = ComponentId.ViewInBrowser;
 			Text = text;